Refactoring feature organization
