+++ src/Orchard.Cms.Web/Modules/Orchard.Body/Startup.cs
@@ -11,7 +11,6 @@ using Orchard.ContentManagement.Handlers;
 using Orchard.ContentTypes.Editors;
 using Orchard.Data.Migration;
 using Orchard.Indexing;
-using Orchard.MetaWeblog;
 using Orchard.Tokens;
 
 namespace Orchard.Body
@@ -29,15 +28,6 @@ namespace Orchard.Body
             services.AddScoped<IContentPartHandler, BodyPartHandler>();
 
             services.AddNullTokenizer();
-        }
-    }
-
-    [OrchardFeature("Orchard.RemotePublishing")]
-    public class RemotePublishingStartup : StartupBase
-    {
-        public override void ConfigureServices(IServiceCollection services)
-        {
-            services.AddScoped<IMetaWeblogDriver, BodyMetaWeblogDriver>();
         }
     }
 }
