+++ src/Orchard.Cms.Web/Modules/Orchard.Lists/RemotePublishing/Startup.cs
@@ -1,0 +1,31 @@
+﻿using System;
+using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Mvc.Modules;
+using Microsoft.AspNetCore.Routing;
+using Microsoft.Extensions.DependencyInjection;
+using Orchard.ContentManagement.Display.ContentDisplay;
+using Orchard.Core.XmlRpc;
+
+namespace Orchard.Lists.RemotePublishing
+{
+
+    [OrchardFeature("Orchard.RemotePublishing")]
+    public class RemotePublishingStartup : StartupBase
+    {
+        public override void ConfigureServices(IServiceCollection services)
+        {
+            services.AddScoped<IXmlRpcHandler, MetaWeblogHandler>();
+            services.AddScoped<IContentPartDisplayDriver, ListMetaWeblogDriver>();
+        }
+
+        public override void Configure(IApplicationBuilder app, IRouteBuilder routes, IServiceProvider serviceProvider)
+        {
+            routes.MapAreaRoute(
+                name: "RSD",
+                areaName: "Orchard.Lists",
+                template: "xmlrpc/metaweblog/{contentItemId}/rsd",
+                defaults: new { controller = "RemotePublishing", action = "Rsd" }
+            );
+        }
+    }
+}
