+++ src/Orchard.Cms.Web/Modules/Orchard.Markdown/Startup.cs
@@ -1,21 +1,20 @@
 ﻿using Microsoft.AspNetCore.Mvc.Modules;
 using Microsoft.Extensions.DependencyInjection;
-using Orchard.Markdown.Drivers;
-using Orchard.Markdown.Handlers;
-using Orchard.Markdown.Indexing;
-using Orchard.Markdown.Model;
-using Orchard.Markdown.Settings;
+using Orchard.ContentFields.Fields;
+using Orchard.ContentFields.Indexing;
+using Orchard.ContentFields.Settings;
 using Orchard.ContentManagement;
 using Orchard.ContentManagement.Display.ContentDisplay;
 using Orchard.ContentManagement.Handlers;
 using Orchard.ContentTypes.Editors;
 using Orchard.Data.Migration;
 using Orchard.Indexing;
+using Orchard.Markdown.Drivers;
+using Orchard.Markdown.Handlers;
+using Orchard.Markdown.Indexing;
+using Orchard.Markdown.Model;
+using Orchard.Markdown.Settings;
 using Orchard.Tokens;
-using Orchard.ContentFields.Fields;
-using Orchard.ContentFields.Settings;
-using Orchard.ContentFields.Indexing;
-using Orchard.MetaWeblog;
 
 namespace Orchard.Markdown
 {
@@ -38,15 +37,6 @@ namespace Orchard.Markdown
             services.AddScoped<IContentFieldIndexHandler, MarkdownFieldIndexHandler>();
 
             services.AddNullTokenizer();
-        }
-    }
-
-    [OrchardFeature("Orchard.RemotePublishing")]
-    public class RemotePublishingStartup : StartupBase
-    {
-        public override void ConfigureServices(IServiceCollection services)
-        {
-            services.AddScoped<IMetaWeblogDriver, MarkdownMetaWeblogDriver>();
         }
     }
 }
