+++ src/Orchard.Cms.Web/Modules/Orchard.Title/RemotePublishing/Startup.cs
@@ -1,0 +1,16 @@
+﻿using Microsoft.AspNetCore.Mvc.Modules;
+using Microsoft.Extensions.DependencyInjection;
+using Orchard.MetaWeblog;
+
+namespace Orchard.Title.RemotePublishing
+{
+
+    [OrchardFeature("Orchard.RemotePublishing")]
+    public class RemotePublishingStartup : StartupBase
+    {
+        public override void ConfigureServices(IServiceCollection services)
+        {
+            services.AddScoped<IMetaWeblogDriver, TitleMetaWeblogDriver>();
+        }
+    }
+}
