+++ src/Orchard.Cms.Web/Modules/Orchard.Title/Startup.cs
@@ -5,7 +5,6 @@ using Orchard.ContentManagement.Display.ContentDisplay;
 using Orchard.ContentManagement.Handlers;
 using Orchard.Data.Migration;
 using Orchard.Indexing;
-using Orchard.MetaWeblog;
 using Orchard.Title.Drivers;
 using Orchard.Title.Handlers;
 using Orchard.Title.Indexing;
@@ -24,15 +23,6 @@ namespace Orchard.Title
             services.AddScoped<IContentPartIndexHandler, TitlePartIndexHandler>();
 
             services.AddScoped<IDataMigration, Migrations>();
-        }
-    }
-
-    [OrchardFeature("Orchard.RemotePublishing")]
-    public class RemotePublishingStartup : StartupBase
-    {
-        public override void ConfigureServices(IServiceCollection services)
-        {
-            services.AddScoped<IMetaWeblogDriver, TitleMetaWeblogDriver>();
         }
     }
 }
